Add application state to crash reports.
Add application state to crash reports by using the crash handler
extension and falling back to the crash log dictionary if the extension
is not supported.

Refactors the h5vcc crash log dictionary singleton to support this.

b/261773222
